fix: share config write lock across mutation routes and use settings.write scope for privacy endpoint
Extract configWriteChain, readConfigFile, writeConfigFileAtomic, and getConfigPath
into a shared config-file-utils module so that concurrent PATCH requests to
/v1/config/privacy and /v1/feature-flags/:key serialize against the same chain,
preventing silent data loss from last-writer-wins races.

Also change the privacy config endpoint scope from feature_flags.write to
settings.write, which is semantically correct for a settings mutation endpoint.
